// SellingCoursesUI/services/interfaces/authenticateInterfaces.tsx
export interface RegisterData {
    firstName: string,
    lastName: string,
    email: string
    sdt: string
    password: string
}
// export interface RegisterData {
//     name: string
//     email: string
//     password: string
// }

export interface LoginData {
    email: string
    password: string
}